requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pillow==10.3.0
open-clip-torch==2.24.0
lightning==2.4.0
